requirements.txt
python-telegram-bot[job-queue,webhooks]==22.6
yfinance==1.2.0
pandas>=2.0
ta==0.11.0
anthropic>=0.84.0
python-dotenv==1.1.0
